+++ ICD.Connect.Calendaring.Robin/Controls/Calendar/RobinBooking.cs
@@ -45,6 +45,9 @@ namespace ICD.Connect.Calendaring.Robin.Controls.Calendar
 			get { return m_Event.IsPrivate.ToLower() == "private"; }
 		}
 
+		public override bool CheckedIn { get { return false; } }
+		public override bool CheckedOut { get { return false; } }
+
 		public override IEnumerable<IDialContext> GetBookingNumbers()
 		{
 			return m_DialContexts.ToArray(m_DialContexts.Count);