+++ ICD.Connect.Calendaring.Robin/Controls/Calendar/RobinServiceDeviceCalendarControl.cs
@@ -91,6 +91,26 @@ namespace ICD.Connect.Calendaring.Robin.Controls.Calendar
 			return m_BookingSection.Execute(() => m_EventToBooking.Values.ToArray(m_EventToBooking.Count));
 		}
 
+		public override bool CanCheckIn(IBooking booking)
+		{
+			return false;
+		}
+
+		public override bool CanCheckOut(IBooking booking)
+		{
+			return false;
+		}
+
+		public override void CheckIn(IBooking booking)
+		{
+			throw new NotSupportedException();
+		}
+
+		public override void CheckOut(IBooking booking)
+		{
+			throw new NotSupportedException();
+		}
+
 		#endregion
 
 		#region Parent Callbacks