+++ ICD.Connect.Calendaring/Controls/AbstractCalendarControl.cs
@@ -41,6 +41,31 @@ namespace ICD.Connect.Calendaring.Controls
 	    [PublicAPI]
 	    public abstract IEnumerable<IBooking> GetBookings();
 
+	    /// <summary>
+	    /// Returns true if the booking argument can be checked in.
+	    /// </summary>
+	    /// <returns></returns>
+	    public abstract bool CanCheckIn(IBooking booking);
+
+	    /// <summary>
+	    /// Returns true if the booking argument can be checked out of.
+	    /// </summary>
+	    /// <param name="booking"></param>
+	    /// <returns></returns>
+	    public abstract bool CanCheckOut(IBooking booking);
+
+	    /// <summary>
+	    /// Checks in to the specified booking.
+	    /// </summary>
+	    /// <param name="booking"></param>
+	    public abstract void CheckIn(IBooking booking);
+
+	    /// <summary>
+	    /// Checks out of the specified booking.
+	    /// </summary>
+	    /// <param name="booking"></param>
+	    public abstract void CheckOut(IBooking booking);
+
 	    #endregion
 
 	    #region Console