+++ ICD.Connect.Calendaring/Controls/ICalendarControl.cs
@@ -21,5 +21,30 @@ namespace ICD.Connect.Calendaring.Controls
 		/// Gets the collection of calendar bookings.
 		/// </summary>
 		IEnumerable<IBooking> GetBookings();
+
+		/// <summary>
+		/// Returns true if the booking argument can be checked in.
+		/// </summary>
+		/// <returns></returns>
+		bool CanCheckIn(IBooking booking);
+
+		/// <summary>
+		/// Returns true if the booking argument can be checked out of.
+		/// </summary>
+		/// <param name="booking"></param>
+		/// <returns></returns>
+		bool CanCheckOut(IBooking booking);
+
+		/// <summary>
+		/// Checks in to the specified booking.
+		/// </summary>
+		/// <param name="booking"></param>
+		void CheckIn(IBooking booking);
+
+		/// <summary>
+		/// Checks out of the specified booking.
+		/// </summary>
+		/// <param name="booking"></param>
+		void CheckOut(IBooking booking);
 	}
 }